Gherkin feature `there should be the following group membership changes:`: - Refactor into a file, because the steps_app_language.go starts to be huge. - Fix `resultCount` should not be 0.
